Move arq client into safir.arq module
This logically separates the FastAPI dependency and its state from the
arq client and domain models.
